Changed entity limits to be based on integers instead of bigintegers

+++ API/src/main/java/com/bgsoftware/superiorskyblock/api/island/algorithms/IslandEntitiesTrackerAlgorithm.java
@@ -2,7 +2,6 @@ package com.bgsoftware.superiorskyblock.api.island.algorithms;
 
 import com.bgsoftware.superiorskyblock.api.key.Key;
 
-import java.math.BigInteger;
 import java.util.Map;
 
 public interface IslandEntitiesTrackerAlgorithm {
@@ -14,7 +13,7 @@ public interface IslandEntitiesTrackerAlgorithm {
      * @param amount The amount of the entity.
      * @return Whether the entity was successfully tracked.
      */
-    boolean trackEntity(Key key, BigInteger amount);
+    boolean trackEntity(Key key, Integer amount);
 
     /**
      * Untrack a entity with a specific amount.
@@ -23,19 +22,19 @@ public interface IslandEntitiesTrackerAlgorithm {
      * @param amount The amount of the entity.
      * @return Whether the entity was successfully untracked.
      */
-    boolean untrackEntity(Key key, BigInteger amount);
+    boolean untrackEntity(Key key, Integer amount);
 
     /**
      * Get the amount of entities that are on the island.
      *
      * @param key The entity's key to check.
      */
-    BigInteger getEntityCount(Key key);
+    Integer getEntityCount(Key key);
 
     /**
      * Get all the entities that are on the island.
      */
-    Map<Key, BigInteger> getEntitiesCounts();
+    Map<Key, Integer> getEntitiesCounts();
 
     /**
      * Clear all the entity counts of the island.
