+++ src/main/java/com/bgsoftware/superiorskyblock/island/algorithms/DefaultIslandEntitiesTrackerAlgorithm.java
@@ -5,15 +5,16 @@ import com.bgsoftware.superiorskyblock.api.island.Island;
 import com.bgsoftware.superiorskyblock.api.island.algorithms.IslandEntitiesTrackerAlgorithm;
 import com.bgsoftware.superiorskyblock.api.key.Key;
 import com.bgsoftware.superiorskyblock.key.dataset.KeyMap;
+import com.bgsoftware.superiorskyblock.utils.islands.IslandUtils;
+import com.google.common.base.Preconditions;
 
-import java.math.BigInteger;
 import java.util.Map;
 
 public final class DefaultIslandEntitiesTrackerAlgorithm implements IslandEntitiesTrackerAlgorithm {
 
     private static final SuperiorSkyblockPlugin plugin = SuperiorSkyblockPlugin.getPlugin();
 
-    private final KeyMap<BigInteger> entityCounts = new KeyMap<>();
+    private final KeyMap<Integer> entityCounts = new KeyMap<>();
 
     private final Island island;
 
@@ -22,22 +23,22 @@ public final class DefaultIslandEntitiesTrackerAlgorithm implements IslandEntiti
     }
 
     @Override
-    public boolean trackEntity(Key key, BigInteger amount) {
+    public boolean trackEntity(Key key, Integer amount) {
         return false;
     }
 
     @Override
-    public boolean untrackEntity(Key key, BigInteger amount) {
+    public boolean untrackEntity(Key key, Integer amount) {
         return false;
     }
 
     @Override
-    public BigInteger getEntityCount(Key key) {
+    public Integer getEntityCount(Key key) {
         return null;
     }
 
     @Override
-    public Map<Key, BigInteger> getEntitiesCounts() {
+    public Map<Key, Integer> getEntitiesCounts() {
         return null;
     }
